+++ lib/nylas/resources/folders.rb
@@ -15,6 +15,10 @@ module Nylas
     #
     # @param identifier [String] Grant ID or email account to query.
     # @param query_params [Hash, nil] Query params to pass to the request.
+    #   Supported parameters include:
+    #   - single_level: (Boolean) For Microsoft accounts only. If true, retrieves folders from
+    #     a single-level hierarchy only. If false (default), retrieves folders across a
+    #     multi-level hierarchy.
     #   - include_hidden_folders [Boolean] (Microsoft only) When true, includes hidden folders.
     # @return [Array(Array(Hash), String, String)] The list of folders, API Request ID, and next cursor.
     def list(identifier:, query_params: nil)