+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### [Unreleased]
+* Added support for `single_level` query parameter in Folders API for Microsoft accounts
 * Added support for `include_hidden_folders` query parameter in folders list endpoint for Microsoft accounts to control whether hidden folders are included in the response
 
 ### 6.5.0 / 2025-06-13